feature: metrics route to allow for Authorization header
